Remove second set of single quotes
The quotes around the regex was not quoting the regex but instead terminating the first quote and starting a new quoted string. This lead to the dollar sign being dropped. The regex actually doesn't need any quoting so they can just be removed.

Also, the dot needs to be escaped.